src/mcp_server/requirements.txt
# ============================================================================
# MCP SERVER - AutoInsight Tab 2
# ============================================================================
certifi
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# Snowflake Connector
snowflake-connector-python==3.5.0

# Environment Variables
python-dotenv==1.0.0

# Dashboard Generation (Claude LLM)
instructor==1.5.0

anthropic>=0.40.0
httpx>=0.27.0,<0.28.0